+++ pkg/util/listtomap/list_to_map.go
@@ -74,14 +74,3 @@ func (l *ListToMap) GetMap() map[string]struct{} {
 	}
 	return *l
 }
-
-func Merge(m1, m2 ListToMap) ListToMap {
-	merged := make(ListToMap)
-	for k := range m1 {
-		merged[k] = struct{}{}
-	}
-	for k := range m2 {
-		merged[k] = struct{}{}
-	}
-	return merged
-}